chore: some log fields comment out, fetch log by date fixed, get logger function created, format code
