+++ app/api/v1/endpoints/user.py
@@ -1,6 +1,6 @@
 import re
-from app.core.logger import LOG
 import app.res.error as http_error
+from app.core.logger import get_logger
 from app.schema.base import BaseResponseDto
 from app.res.openapi_error import UserResponseDoc
 from fastapi import APIRouter, HTTPException, Query
@@ -12,7 +12,9 @@ user_router = APIRouter()
 
 @user_router.get('/fetch')
 async def fetch_users() -> UsersDto:
+    logger = get_logger()
     try:
+        
         users = [
             {
                 "user_id": "2121",
@@ -25,43 +27,45 @@ async def fetch_users() -> UsersDto:
                 "user_type": "SELLER"
             }
         ]
-        LOG.debug("user list", extra={"user list": users})
+        logger.debug("user list", extra={"obj": users})
 
         return UsersDto(code="SUCCESS", message="Success", users=users)
 
     except HTTPException as ex:
-        LOG.error(f"HTTP Exception: {ex.detail}")
+        logger.error(f"HTTP Exception: {ex.detail}")
         raise ex
     except Exception as ex:
-        LOG.error(f"Unexpected error occurred: {ex}", exc_info=True)
+        logger.error(f"Unexpected error occurred: {ex}", exc_info=True)
         raise http_error.InternalServerError()
     
 
 @user_router.post('/create', responses=UserResponseDoc.create)
 async def create_user(req: CreateUserDto) -> BaseResponseDto:
+    logger = get_logger(user_id=req.user_id)
     try:
-
+        
         try:
             if re.search(r"[^a-zA-Z0-9_]", req.name):
                 raise http_error.UserNameInvalid()
-            LOG.info(f"Create user {req.name}")
+            logger.info(f"Create user {req.name}")
         except ValueError:
             raise http_error.UserIDAlreadyExist()
         
 
-        LOG.debug("User created")
+        logger.debug("User created")
         return BaseResponseDto(code="SUCCESS", message="User Successfully Created")
     
     except HTTPException as ex:
-        LOG.error(f"HTTP Exception: {ex.detail}")
+        logger.error(f"HTTP Exception: {ex.detail}")
         raise ex
     except Exception as ex:
-        LOG.error(f"Unexpected error occurred: {ex}", exc_info=True)
+        logger.error(f"Unexpected error occurred: {ex}", exc_info=True)
         raise http_error.InternalServerError()
 
 
 @user_router.delete('/delete', responses=UserResponseDoc.delete)
 async def delete_user(user_id: str = Query(..., max_length=10, description="user id assignment")) -> BaseResponseDto:
+    logger = get_logger(user_id=user_id)
     try:
 
         user = {
@@ -71,13 +75,13 @@ async def delete_user(user_id: str = Query(..., max_length=10, description="user
         if user is None:
             raise http_error.UserNotExist()
         
-        LOG.debug("User deleted")
+        logger.debug("User deleted")
 
         return BaseResponseDto(code="SUCCESS", message="User Successfully Deleted")
     
     except HTTPException as ex:
-        LOG.error(f"HTTP Exception: {ex.detail}")
+        logger.error(f"HTTP Exception: {ex.detail}")
         raise ex
     except Exception as ex:
-        LOG.error(f"Unexpected error occurred: {ex}", exc_info=True)
+        logger.error(f"Unexpected error occurred: {ex}", exc_info=True)
         raise http_error.InternalServerError()
