+++ app/assets/logger_view.html
@@ -44,14 +44,12 @@
         }
 
         function fetchLogs() {
-            const dateInput = document.getElementById("logDate").value;
-            if (!dateInput) {
+            const selectedDate = document.getElementById("logDate").value;
+            if (!selectedDate) {
                 alert("Please select a date.");
                 return;
             }
-            const selectedDate = new Date(dateInput).toISOString().slice(0, 10);
-            const currentDate = new Date().toISOString().slice(0, 10);
-            const logDate = (selectedDate === currentDate) ? "today" : dateInput.replace(/-/g, '');
+            const logDate = selectedDate.replace(/-/g, '');
             // Get the full current URL path dynamically
             const fullPath = window.location.href.split('?')[0];
             // Construct the URL by replacing any existing date or query part
@@ -69,8 +67,9 @@
         <p>Contact Email: <a href='mailto:omer.anzar2@gmail.com'
                 class='text-blue-400 hover:underline'>omer.anzar2@gmail.com</a></p>
         <p>GitHub Repository: <a href='https://github.com/Hoopoes/FastStart' target='_blank'
-                class='text-blue-400 hover:underline'>FastStart on GitHub</a></p>
+                class='text-blue-400 hover:underline'>FastStart</a></p>
     </header>
+    </br>
     <div class='bg-gray-800 shadow-md rounded-lg p-4 mb-6'>
         <div class='flex justify-between items-center'>
             <h1 class='text-2xl font-bold text-white'>Logs for ${date}</h1>
@@ -78,7 +77,7 @@
                 <input type='date' id='logDate' class='p-2 bg-gray-700 text-gray-300 border border-gray-600 rounded'>
                 <button onclick='fetchLogs()'
                     class='flex items-center gap-1 p-2 bg-blue-600 text-white rounded hover:bg-blue-700'>
-                    <span class="material-icons">search</span> Search
+                    <span class="material-icons">folder_open</span> Browse
                 </button>
             </div>
         </div>
